Extract createFeatureGeometrySlice from projectStore (P5 store slice)
